Fix event always being created error
This commit fixes a bug where the event editor always creates an event
rather than updating them. The issue was that the event_id which was used to
determine if you are creating or updating was initially
stored in the event editor and initialized to -1, but then not set to
the appropriate value later on.