Event Management Functionality (#116)
* Implement event editor

This commit implements the event editor, which allows admin users to
create and update events. There are currently no buttons yet to access
this page without manually going to the route.

* Fix Existing Errors From Merge

* Create Event List Widget in Shared Module

This commit creates the Event List widget in the Shared Module so that it can be accessed from both events and organizations.

* Major Refactor + Add Events Lists to Organization Details

This commit is a major refactor that fixes some lingering errors in the backend. In addition, the events list now shows on the organization detail page.

* Add event creation button

This commit adds and implements the Event creation button so users may
add events to orgaizations they have permissions for.

* organization.e

* Make event management buttons appear on Events Page

This commit modifies the code so that the event management buttons
appear on both the Events Page and the Events Details Page. This allows
for easier access to event management buttons.

* Add confirmation snack bar for event deletion

This commit adds a confirmation snack bar to ensure that the user
actually wants to delete the event before sending the delete HTTP
request.

* Fix event always being created error

This commit fixes a bug where the event editor always creates an event
rather than updating them. The issue was that the event_id which was used to
determine if you are creating or updating was initially
stored in the event editor and initialized to -1, but then not set to
the appropriate value later on.

* Fix Event List Padding on Events Page

* Fix Event Creation Issue, Add Documentation

---------

Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>