Add a document explaining/listing feature permissions
This commit adds a document to the docs directory that defines the
standard permissions used in the backend to authorize users for feature
actions. This ensures that permissions are added properly in the
frontend by providing a reference file for existing action/resource
sets.

Permissions for coworking still need to be defined!