Fix Organization Permissions (#157)
Fix backend organization exceptions and permissions

Updates the Organization Service exceptions to follow the
conventions in the other services, particularly in terms of using
the ResourceNotFoundException and removing unnecessary else statements.

Updates the permission check for the create/update/delete
API routes to be accurate and specific to the action.

Updates the Organization Details frontend, its
corresponding widgets, and the organization editor to enforce the
appropriate permissions for creating and deleting events.

Fixes some linting errors in the HTML.

Update admin permissions doc according to changes
This commit updates the admin permissions doc to reflect the changes
with permissions. Namely, instead of using the organization_id when
granting the "update" action permission, we use the organization_slug.
This is primarily because the canActivateFn used when checking the
user permissions only has access to the path parameters, which does not
include the organization_id.