Add oc OpenShift Client to DevContainer
DevContainer Dockerfile now downloads and installs the oc client for
OpenShift. This enables a workflow of working with staging/production
within the DevContainer terminal and removes a small hurdle for students
needing to download and correctly install oc. Even more optimistically
this enables us to begin scripting many devops/deployment concerns in
the codebase. Some thought will need to be put into guard rails to avoid
accidentally scripting against the production site before scripting
anything, though!

As this changes the dev Dockerfile, a rebuild of the DevContainer is
needed once the commit is pulled into a dev repository.