Add Display Bearer Token feature to profile view (#186)
Users can now display their bearer token from the CSXL website by viewing their profile page. The ability to quickly copy without viewing, in a single click, is also added. The value of this feature is simplifying the developer experience when working with the API which requires the bearer token for authentication.

---------

Co-authored-by: Kris Jordan <kris@cs.unc.edu>