Implement event register/unregister button This commit implements the event register/unregister button on the event detail card.
In order to do so, it adds subject and isRegistered$
fields to the Event Service to ensure that the detail card updates with
the appropriate button depending on if the user is registered for the
event or not.

I also added a new API call to the backend to directly get retrieve the
registration status of a user for an event as a boolean value, rather
than using the "getEventRegistration" request and piping the result to be
a boolean.

More work needs to be done to prevent duplicate API calls to retrieve
registration statuses for events on the Event Page.