Script migration: backend.script.generate_migration
The `backend.script.generate_migration` script can be used to generate
migrations. It takes two arguments: the name of the remote with the
feature branch PR the migration is being generated for and the name of
the branch.

The script then follows a sequence of steps to generate an alembic
migration based off of the state of the `main` branch.