Remove can_register Column from Event Entity #232 (#235)
* Remove organizers from paginated list of registrations

* Fix Github avatars and improve display of organizers
This commit updates the backend EventOrganizer model to include the
Github avatar of the user and likewise updates the frontend to ensure
that the desired user avatar is displayed rather than the currently
logged in user's avatar.

This commit also removes organizer registrations from the list of
registrations. Instead, organizers are displayed on the event details
card. Clicking the organizer chip will allow you to email that organizer.

* Remove can_register column from Event Entity
This commit removes the can_register column from the Event Entity as it
is simply a computer value based on registration_limit being greater
than 0. It likewise updates all instances of can_register and either
removes it or replaces the logic to just use registration_limit.

* Generate migration script
This commit generates the migration script for merging in to main using
the migration generate script.

* Clean up migration script file

* Fix failing event service tests

---------

Co-authored-by: Kris Jordan <kris@cs.unc.edu>