Convert Reusable Code to Widgets & Refactor Event/Organizer Models (#237)
* Convert event editor user lookup to a reusable widget
This converts the event editor user lookup form field to a widget so it
can be used throughout the application.

Additional Changes to Be Made: Create a new model to represent public
users/info so the form field can be used for ANY user.

* Create new `PublicUser` model
This commit creates a PublicUser model to represent public information
about a user. This is surfaced on the frontend as "PublicProfile" to
match referencing the User as "Profile".

This PublicUser model replaces the EventOrganizer and EventMember models.

* Remove attendees from Event model
This commit removes the list of attendees from the Event model to avoid
exposing the attendee information when retrieving events. This list is
never actually used, so it does not impact the code.

* Create user chip list widget
This commit turns the user chip list on the event details card into a
widget so it can be reused to display user information. It also adds
an "enableMailTo" field so users may be listed without allowing other
users to email them.