230 click  pointer affordance of academics admin tables (#245)
* Move (click)="updateCourse(elt)" from p tag to div tag

Before, the admin was only able to edit the desired course if they
clicked on the actual text. This change allows the admin to click
anywhere on the row and get redirected to the edit course page.

* Update deleteCourse to work with change

After makinig the change in my previous commit, the page would
redirect to the edit course page when clicking on delete. To fix, I
modified the method to take in an event and call the
stopPropogation method in order to prevent any parent event
handlers from being executed.

* update documentation

* same changes that were made to course.component html and ts

* fixed typo

* fixed typo

* allow admin to click anywhere in row instead of just the text

* add event to delete function to stop propagation

* allow user to click anywhere in the row instead of just the text

* add event to delete method to stop propagation