Remove absolute from Card Actions
When condensing the card, there is an auto sizing for each media query.
This caused an unintended consequence of superimposing the action buttons on the text details (Who, When, Where).
The fix was to remove the absolute positioning in favor of relative position.
This corrected the space of the buttons on the card.