Implement User Reservation Hourly Limit
I added a helper method that will check if a user has gone over their
weekly alloted time of 6 hours per week. The helper method returns
a boolean value that will get dealt with in the draft reservation method.
Note: this method does not count any reservation for the CSXL
Also, I added a method that computes the current amount of hours that
a user has reserved for the week. I still need to implement this on the
front end