Integrate Event Pagination (#276)
* Add pagination for upcoming events

* Add past events button

Still need to work on sorting events passed in from backend (currently clashing with frontend sorting). Button should also change to "Show current events" after pressing. Or have a checkbox thing.

* Finish pagination feature

Further considerations:
Adding query params to URL for sharing
I'm thinking we can get rid of the evetnFilterPipe, want to confirm

* Remove EventFilterPipe

Searching is now done on the backend based on parameters passed to the HTTP request, so I am removing the frontend filter.

* Incorporate date range into URL query parameters

For linking to a specific page of events

* Add debouncing to search bar

* Refactor to fix edge case

Event list was not refreshing when navigating back to the events tab. Now the page checks for a change in the route to update the event list

* Remove unecessary parameters

* Add new demo data and adjust search

* Remove extra lines

* Make search sort consistent

* Change en-GB to en-US

* Update Event Pagination imports

* Update paginated events service method

* Improve event page/pagination UI

* Fix column alignment

* Fix search bar

* Ensure events are visible to unauthenticated users
This commit updates the backend code and frontend code
for the event page to ensure that paginated events are
visible to unauthenticated users.

* Add test for retrieving paginated events when unauthenticated

* Fix documentation and event page margins

---------

Co-authored-by: tdianaa <rtoneyd@gmail.com>
Co-authored-by: Kris Jordan <krisjordan@gmail.com>