Create Walk-in Reservations for XL Members from Ambassador Welcome Desk (#313)
Ambassadors requested the ability to create a drop-in reservation for a community member showing up to the XL without having already created a reservation.

This commit adds a new area to the Ambassador Home Component that allows searching for a community member and registering an available seat for them. Since this is staff-only, the UX around pop-up notifications is just alerts for now. Future work could improve this to make use of snackbar or other material UX.

To implement this feature, a few other useful additions were made in other parts of the application, including:

Search for XL members by partial PID (with unit test coverage)
Get notified of changes from the UsersLookup widget so that components which use it can have callback methods (example use case shown in this widget: after a user is selected, we lookup available seats and show the control for selecting a seat type)