Room Reservations (#250)
This very significant feature landing adds the ability to reserve group rooms in the XL into the future. Full-stack implications from reservation UI to resource limitations on the backend. Feature implemented by group B8 in the Fall of 2023 and Spring of 2024. For a full run-down of this feature, see `docs/specs/room-reservation.md`.