Fix front-end issue with draft drop-in reservation (#356)
Due to some changes that merged with room reservations, when creating a
drop-in reservation and navigating back (intentionally or not) to home
you would not see the draft reservation anymore because it was filtered
from the home page. This meant a draft reservation was created but could
not be cancelled by user or ambassador. This PR fixes this issue by
making the draft of a drop-in visible on the home page. No changes in
behavior for room reservations.