Fix bug with end_idx logic for future reservations
The current problem with the code is that at 11 pm, we were
not able to see future reservations in red for tomorrow. This was
actually a red herring that led me to look at parts of the code
that were not responsible for this.

Then suddenly at 12 am, everything started working as intended.
This was the biggest hint for me in debugging this.

Basically on line 345, I just had to add in a condition to check
that the date I am getting as a param is the same date as today.
Only in that case, should I execute the logic of
end_idx < current_time_idx. However, earlier this was being
executed for all reservations which is why whenever we had
any reservations which had an end time of less than 11 pm
(which most reservations do), our function would just skip over
them. This also aligns with what we saw at 12 am when
everything seemed to work as intended. This is again because
most reservations don't have an ending time before 12 am.

Anyway, took me over 60 minutes to add in these 30 chars to fix
the bug. So I typed an average of 0.5 char per minute... :,)