Routed TvComponent around Navigation Component
Observed any changes to the route for when the routing ends so we can
work with a completed and resolved route instead of a snapshot, which
only contains the route at the `app.component` level. This is then
subscribed to, the correct route information is extracted, and tested
in the `app.component.html` page.