Get test running
Fix required changing the order of some fixture imports and adding more
imports that previous imports were dependant upon.

This seems very clunky and it would make more sense if the fixtures to
insert data required for say the seats were imported in the seat_data
file so we wouldn't have to worry about importing those in every test
file.