Feature/join organization (#26)
* Updated roster styling and membership role display

* Refactor API into separate organization roster service

Instead of placing organization roster management in the same code as organization details, it will be moved into its own service and injected in separately. This makes it easier to modify and expand to add filtering/editing logic and decouples it from the responsibility of fetching org metadata, which is fairly static and unrelated.

This involves refactoring roster API methods into a separate OrganizationRoster service file.

The details widget now passes org/user info and the service into the roster so that it can filter/perform operations within itself.

The roster must still be fetched in the parent in case we eventually want to display membership on the details card or not render the roster for empty/non-rostered clubs (CS Dept). The service must also be present so we can join from the details page.

* Add POST membership API method

Fix backend API route to not include member_id

* Add join button linked to API calls

Add a join button that when clicked will add the current user to the roster. This does not yet refresh the view.

* Add Body to backend imports to support request body parameters

* Add error-trapping for add and remove member endpoints

* Test all organization membership service functions

* Add temporary user-facing error to frontend for joining an already joined organization

---------

Co-authored-by: anikaahmed114 <aahm@unc.edu>
Co-authored-by: Alanna423 <alanna.zhang423@gmail.com>