OH Changes to Support Quick Refreshes
These changes are to support quick refreshes that will occur when
fetching fast data. I want to make sure that we can quickly update
the queue number without needing to completly referesh the widget, which
we will only do now when we notice a difference in office hours being
offered.

Some of these include only running the onChanges when we notice a change
other than the queued number and switching everything to referencing off
of the OfficeHours list.