Finish OfficeHoursWidget with Lazy Loading
Pagination was implemented so that the office hours pages are cycled
from left to right. Allowing the user to have more time to view one
column.

Lazy loading was also implemented to make sure we are changing the
order of columns in time with the spinner widget. However, if only
the queue lengths of all the office hours are changed, then this is
ignored and the update is sent right through.