OH Sounds (#699)
* Update testing data's term to Spring 2025

The variable for Spring 2025's term was already present, I simply
changed its name to match the format for other terms and inserted it
where necessary (terms, current_term).

* Add handleNotification to polling for office hours

The function handleNotification has been added to send a notification
to the TA or student based on certain conditions. The notification will
consist of a chime being played and the title flashing for a short
duration.

* Revert backend test data changes

Due to the scope of changing term data in a way that passes all tests,
the backend changes are reverted in this commit. Future changes to this
data and the corresponding tests will be in their own branch.

* Apply style changes to constants, if statement

The notification titles and sound have been pulled out of the component
as capitalized constants, and two if statements have been combined into
one since they were unnecessarily separated.

* Update documentation for notify variable