build!: upgrade to angular v20 and remove vulnerabilities (#809)
* build: fix critical node vulnerabilities

Fixes the critical node vulnerabilities for security reasons.

* build: upgrade to angular v19

Upgrades the application to Angular v19 and changes the build process
to use the new recommended defaults.

* build: upgrade to angular material v19

Upgrades the app to use Angular Material v19.

* build: upgrade to angular v20

Run the migration to upgrade to Angular v20.

* build: upgrade to Angular Material v20

Runs the migration to upgrade to Angular Material v20.

* build: upgrade angular youtube player to v20

Upgrades the Angular YouTube player dependency to v20.